+++ inte_ae/pdf_reports/inte_pdf_report_mixin.py
@@ -7,9 +7,11 @@ class InteCrfReportMixin:
 
     @property
     def unblinded(self):
-        UnblindingRequest = django_apps.get_model("inte_prn.unblindingrequest")
+        unblinding_request_model_cls = django_apps.get_model(
+            "inte_prn.unblindingrequest"
+        )
         try:
-            unblinded = UnblindingRequest.objects.get(
+            unblinded = unblinding_request_model_cls.objects.get(
                 subject_identifier=self.subject_identifier, approved=True
             )
         except ObjectDoesNotExist: