+++ inte_consent/apps.py
@@ -6,9 +6,3 @@ class AppConfig(DjangoAppConfig):
     verbose_name = "INTE: Consent"
     include_in_administration_section = True
     has_exportable_data = True
-
-    def ready(self):
-        from .models.signals import (  # noqa
-            subject_consent_on_post_save,  # noqa
-            subject_consent_on_post_delete,  # noqa
-        )  # noqa