+++ inte_edc/settings/defaults.py
@@ -2,10 +2,8 @@ import environ
 import os
 import sys
 
-from django.core.exceptions import ImproperlyConfigured
-from edc_sites import get_site_id
+from edc_sites import get_site_from_environment
 from edc_utils import get_datetime_from_env
-from inte_sites import inte_sites
 from pathlib import Path
 
 
@@ -17,15 +15,8 @@ class DisableMigrations:
         return None
 
 
-# simple version check
-try:
-    assert (3, 7) <= (sys.version_info.major, sys.version_info.minor) <= (3, 8)
-except AssertionError:
-    raise ImproperlyConfigured(
-        "Incorrect python version. Expected 3.6 or 3.7. Check your environment."
-    )
-
-BASE_DIR = str(Path(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
+BASE_DIR = str(Path(os.path.dirname(os.path.abspath(__file__))).parent.parent)
+ENV_DIR = str(Path(os.path.dirname(os.path.abspath(__file__))).parent.parent)
 
 env = environ.Env(
     AWS_ENABLED=(bool, False),
@@ -52,18 +43,28 @@ env = environ.Env(
     TWILIO_ENABLED=(bool, False),
 )
 
-# copy your .env file from .envs/ to BASE_DIR
+# copy your .env file from .envs/ to ENV_DIR
 if "runtests.py" in sys.argv:
-    env.read_env(os.path.join(BASE_DIR, ".env-tests"))
-    print(f"Reading env from {os.path.join(BASE_DIR, '.env-tests')}")
+    env.read_env(os.path.join(ENV_DIR, ".env-tests"))
+    print(f"Reading env from {os.path.join(ENV_DIR, '.env-tests')}")
 else:
-    env.read_env(os.path.join(BASE_DIR, ".env"))
+    env.read_env(os.path.join(ENV_DIR, ".env"))
 
 DEBUG = env("DJANGO_DEBUG")
 
 SECRET_KEY = env.str("DJANGO_SECRET_KEY")
 
 APP_NAME = env.str("DJANGO_APP_NAME")
+
+# extract country and sitename from DJANGO_SETTINGS_MODULE environment variable
+EDC_SITES_MODULE_NAME = env.str("EDC_SITES_MODULE_NAME")
+COUNTRY, SITE_ID, _ = get_site_from_environment(
+    default_site_name="kinoni",
+    default_country="uganda",
+    app_name=APP_NAME,
+    sites_module_name=EDC_SITES_MODULE_NAME,
+)
+
 
 LIVE_SYSTEM = env.str("DJANGO_LIVE_SYSTEM")
 
@@ -108,13 +109,16 @@ INSTALLED_APPS = [
     "edc_visit_schedule.apps.AppConfig",
     "edc_dashboard.apps.AppConfig",
     "edc_data_manager.apps.AppConfig",
+    "edc_device.apps.AppConfig",
     "edc_export.apps.AppConfig",
     "edc_fieldsets.apps.AppConfig",
     "edc_form_validators.apps.AppConfig",
+    "edc_identifier.apps.AppConfig",
     "edc_lab_dashboard.apps.AppConfig",
     "edc_label.apps.AppConfig",
     "edc_list_data.apps.AppConfig",
     "edc_locator.apps.AppConfig",
+    "edc_metadata.apps.AppConfig",
     "edc_metadata_rules.apps.AppConfig",
     "edc_model_admin.apps.AppConfig",
     "edc_navbar.apps.AppConfig",
@@ -122,6 +126,7 @@ INSTALLED_APPS = [
     "edc_offstudy.apps.AppConfig",
     "edc_pharmacy.apps.AppConfig",
     "edc_pdutils.apps.AppConfig",
+    "edc_protocol.apps.AppConfig",
     "edc_prn.apps.AppConfig",
     "edc_randomization.apps.AppConfig",
     "edc_reference.apps.AppConfig",
@@ -132,6 +137,7 @@ INSTALLED_APPS = [
     "edc_sites.apps.AppConfig",
     "edc_subject_dashboard.apps.AppConfig",
     "edc_timepoint.apps.AppConfig",
+    "edc_visit_tracking.apps.AppConfig",
     "inte_consent.apps.AppConfig",
     "inte_lists.apps.AppConfig",
     "inte_dashboard.apps.AppConfig",
@@ -147,12 +153,6 @@ INSTALLED_APPS = [
     "inte_export.apps.AppConfig",
     "inte_screening.apps.AppConfig",
     "inte_sites.apps.AppConfig",
-    # "inte_edc.apps.EdcAppointmentAppConfig",
-    "inte_edc.apps.EdcDeviceAppConfig",
-    "inte_edc.apps.EdcIdentifierAppConfig",
-    "inte_edc.apps.EdcMetadataAppConfig",
-    "inte_edc.apps.EdcProtocolAppConfig",
-    "inte_edc.apps.EdcVisitTrackingAppConfig",
     "inte_edc.apps.EdcFacilityAppConfig",
     "inte_edc.apps.AppConfig",
 ]
@@ -411,7 +411,7 @@ EDC_RANDOMIZATION_UNBLINDED_USERS = env.list("EDC_RANDOMIZATION_UNBLINDED_USERS"
 EDC_RANDOMIZATION_REGISTER_DEFAULT_RANDOMIZER = env(
     "EDC_RANDOMIZATION_REGISTER_DEFAULT_RANDOMIZER"
 )
-
+EDC_RANDOMIZATION_SKIP_VERIFY_CHECKS = True
 # django-simple-history
 SIMPLE_HISTORY_REVERT_ENABLED = False
 
@@ -451,15 +451,6 @@ else:
     if env("DJANGO_LOGGING_ENABLED"):
         from .logging.standard import LOGGING  # noqa
 
-# read SITE specific variables from separate .env file
-env.read_env(os.path.join(BASE_DIR, ".env-site"))
-COUNTRY = env.str("DJANGO_COUNTRY")
-# get site ID from more familiar town name
-DJANGO_SITE_NAME = env.str("DJANGO_SITE_NAME")
-if DJANGO_SITE_NAME:
-    SITE_ID = get_site_id(DJANGO_SITE_NAME, sites=inte_sites)
-else:
-    SITE_ID = env.int("DJANGO_SITE_ID")
 
 # if running tests ...
 if "test" in sys.argv or "runtests" in sys.argv: