+++ inte_edc/wsgi.py
@@ -2,6 +2,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "inte_edc.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "inte_edc.settings.defaults")
 
 application = get_wsgi_application()