+++ inte_form_validators/models.py
@@ -1,4 +1,3 @@
-from django.db import models
 from edc_model.models import BaseUuidModel