+++ inte_sites/sites.py
@@ -1,20 +1,28 @@
+from django.core.management import color_style
+
+style = color_style()
+
 fqdn = "inte.clinicedc.org"
 
-inte_sites = (
-    (1, "reviewer", ""),
-    (100, "kinoni", "Kinoni"),
-    (110, "bugamba", "Bugamba"),
-    (120, "bwizibwera", "Bwizibwera"),
-    (140, "ruhoko", "Ruhoko"),
-    (150, "kyazanga", "Kyazanga"),
-    (160, "bukulula", "Bukulula"),
-    (170, "kojja", "Kojja"),
-    (180, "mpigi", "Mpigi"),
-    (190, "namayumba", "Namayumba"),
-    (200, "buwambo", "Buwambo"),
-    (210, "kajjansi", "Kajjansi"),
-    (220, "tikalu", "Tikalu"),
-    (230, "namulonge", "Namulonge"),
-    (240, "kasanje", "Kasanje"),
-    (250, "kasangati", "Kasangati"),
-)
+all_sites = {
+    "uganda": (
+        (101, "bugamba", "Bugamba"),
+        (102, "bwizibwera", "Bwizibwera"),
+        (103, "kinoni", "Kinoni"),
+        (104, "ruhoko", "Ruhoko"),
+        (105, "kyazanga", "Kyazanga"),
+        (106, "bukulula", "Bukulula"),
+        (107, "kojja", "Kojja"),
+        (108, "mpigi", "Mpigi"),
+        (109, "namayumba", "Namayumba"),
+        (110, "buwambo", "Buwambo"),
+        (111, "kajjansi", "Kajjansi"),
+        (112, "tikalu", "Tikalu"),
+        (113, "namulonge", "Namulonge"),
+        (114, "kasanje", "Kasanje"),
+        (115, "kasangati", "Kasangati"),
+        (116, "sekiwunga", "Sekiwunga"),
+        (117, "muduma", "Muduma"),
+        (118, "kyamulibwa", "Kyamulibwa"),
+    )
+}