+++ tests/inte_test_case_mixin.py
@@ -1,3 +1,6 @@
+import string
+from random import choices
+
 from dateutil.relativedelta import relativedelta
 from django.contrib.auth.models import Group
 from django.contrib.sites.models import Site
@@ -9,16 +12,16 @@ from edc_facility.import_holidays import import_holidays
 from edc_facility.models import Holiday
 from edc_list_data.site_list_data import site_list_data
 from edc_randomization.randomization_list_importer import RandomizationListImporter
-from edc_sites import add_or_update_django_sites
+from edc_sites import add_or_update_django_sites, get_sites_by_country
 from edc_sites.tests.site_test_case_mixin import SiteTestCaseMixin
 from edc_utils.date import get_utcnow
 from edc_visit_schedule.constants import DAY1
 from edc_visit_tracking.constants import SCHEDULED
 from inte_auth.codenames_by_group import get_codenames_by_group
-from inte_screening.constants import NCD_CLINIC, HIV_CLINIC
+from inte_screening.constants import HIV_CLINIC
 from inte_screening.forms import SubjectScreeningForm
 from inte_screening.models import SubjectScreening
-from inte_sites.sites import fqdn, inte_sites
+from inte_sites.sites import fqdn
 from inte_subject.models import SubjectVisit
 from model_bakery import baker
 
@@ -26,7 +29,7 @@ from model_bakery import baker
 class InteTestCaseMixin(SiteTestCaseMixin):
     fqdn = fqdn
 
-    default_sites = inte_sites
+    default_sites = get_sites_by_country("uganda")
 
     site_names = [s[1] for s in default_sites]
 
@@ -36,7 +39,7 @@ class InteTestCaseMixin(SiteTestCaseMixin):
     def setUpClass(cls):
         super().setUpClass()
         import_holidays(test=True)
-        add_or_update_django_sites(sites=inte_sites, fqdn=fqdn)
+        add_or_update_django_sites(sites=get_sites_by_country("uganda"), fqdn=fqdn)
         site_list_data.autodiscover()
         GroupPermissionsUpdater(
             codenames_by_group=get_codenames_by_group(), verbose=True
@@ -61,14 +64,16 @@ class InteTestCaseMixin(SiteTestCaseMixin):
                 user.groups.add(group)
         return self.client.force_login(user or self.user)
 
-    def get_subject_screening(self, report_datetime=None, eligibility_datetime=None, **kwargs):
+    def get_subject_screening(
+        self, report_datetime=None, eligibility_datetime=None, **kwargs
+    ):
         data = {
             "screening_consent": YES,
             "age_in_years": 25,
             "clinic_type": HIV_CLINIC,
             "gender": MALE,
             "hospital_identifier": "13343322",
-            "initials": "EW",
+            "initials": "".join(choices(string.ascii_uppercase, k=2)),
             "lives_nearby": YES,
             "qualifying_condition": YES,
             "report_datetime": report_datetime or get_utcnow(),
@@ -101,14 +106,16 @@ class InteTestCaseMixin(SiteTestCaseMixin):
             screening_identifier=subject_screening.screening_identifier,
             initials=subject_screening.initials,
             dob=get_utcnow().date()
-                - relativedelta(years=subject_screening.age_in_years),
+            - relativedelta(years=subject_screening.age_in_years),
             site=Site.objects.get(name=site_name),
             clinic_type=HIV_CLINIC,
         )
         options.update(**kwargs)
         return baker.make_recipe("inte_consent.subjectconsent", **options)
 
-    def get_subject_visit(self, visit_code=None, subject_screening=None, subject_consent=None):
+    def get_subject_visit(
+        self, visit_code=None, subject_screening=None, subject_consent=None
+    ):
         visit_code = visit_code or DAY1
         subject_screening = subject_screening or self.get_subject_screening()
         subject_consent = subject_consent or self.get_subject_consent(subject_screening)