+++ tests/test_labs.py
@@ -16,15 +16,13 @@ class TestLabs(TestCase):
 
     def test_lab_profile_model(self):
         obj = site_labs.get(lab_profile_name="subject_lab_profile")
-        self.assertEqual("inte_subject.subjectrequisition",
-                         obj.requisition_model)
+        self.assertEqual("inte_subject.subjectrequisition", obj.requisition_model)
 
     def test_panel_model(self):
         for panel in site_labs.get(
-                lab_profile_name="subject_lab_profile"
+            lab_profile_name="subject_lab_profile"
         ).panels.values():
-            self.assertEqual(panel.requisition_model,
-                             "inte_subject.subjectrequisition")
+            self.assertEqual(panel.requisition_model, "inte_subject.subjectrequisition")
 
     def test_panels_exist(self):
         self.assertGreater(Panel.objects.all().count(), 0)