+++ tests/test_reportables.py
@@ -1,8 +1,7 @@
-from unittest import skip
-
 from django.test import TestCase, tag
 from edc_reportable import site_reportables, ParserError
 from tempfile import mkdtemp
+from unittest import skip
 
 
 @skip