+++ tests/test_reports.py
@@ -1,11 +1,10 @@
-from unittest import skip
-
 from django.contrib.auth.models import User
 from django.test import TestCase, tag
 from django.test.client import RequestFactory
 from edc_adverse_event.models import AeClassification
 from inte_reports.ae_report import AeReport
 from model_bakery import baker
+from unittest import skip
 
 from .inte_test_case_mixin import InteTestCaseMixin